Remove namespace from presenter & spec
